feat(FN-3675): move helpers to common and send emails
