+++ dtfs-central-api/src/constants/email-template-ids.ts
@@ -1,5 +1,13 @@
 const REPORT_RECONCILED = 'f6a10d2f-a1bb-4640-9bf0-0699ef442551';
 const FEE_RECORD_CORRECTION_REQUEST = 'c32528aa-4e12-43c1-9ff0-11f327513b9e';
+const FEE_RECORD_CORRECTION_RECEIVED_BANK_NOTIFICATION = 'c60489c5-a2e8-4100-a63a-3f5e691bdb20';
+const FEE_RECORD_CORRECTION_RECEIVED_UKEF_NOTIFICATION = '2af491e6-2968-4476-8db9-74202138dfad';
 const REPORT_INSERTION_CRON_FAILURE = 'a8a2b0db-67a8-409b-b8b7-7104b137acfd';
 
-export default { REPORT_RECONCILED, FEE_RECORD_CORRECTION_REQUEST, REPORT_INSERTION_CRON_FAILURE };
+export default {
+  REPORT_RECONCILED,
+  FEE_RECORD_CORRECTION_REQUEST,
+  FEE_RECORD_CORRECTION_RECEIVED_BANK_NOTIFICATION,
+  FEE_RECORD_CORRECTION_RECEIVED_UKEF_NOTIFICATION,
+  REPORT_INSERTION_CRON_FAILURE,
+};
