+++ libs/common/src/helpers/index.ts
@@ -21,3 +21,4 @@ export * from './is-ukef-email';
 export * from './fee-record-corrections';
 export * from './decode-html-entities';
 export * from './map-currencies-to-radio-items';
+export * from './utilisation-report-emails';
