+++ portal-api/src/v1/services/utilisation-report/email-service.ts
@@ -1,35 +1,7 @@
-import z from 'zod';
-import { InvalidEnvironmentVariableError, PaymentOfficerTeam, asString, formatDateForEmail } from '@ukef/dtfs2-common';
+import { PaymentOfficerTeam, formatDateForEmail, getTfmUiUrl, getUkefGefReportingEmailRecipients } from '@ukef/dtfs2-common';
 import sendEmail from '../../email';
 import { EMAIL_TEMPLATE_IDS } from '../../../constants';
 import api from '../../api';
-
-const getUkefGefReportingEmailRecipients = () => {
-  const EmailsSchema = z.array(z.string().email());
-
-  try {
-    const emails = EmailsSchema.parse(JSON.parse(asString(process.env.UKEF_GEF_REPORTING_EMAIL_RECIPIENT, 'UKEF_GEF_REPORTING_EMAIL_RECIPIENT')));
-    return emails;
-  } catch (error) {
-    console.error('Failed to parse UKEF_GEF_REPORTING_EMAIL_RECIPIENT ', error);
-    throw new InvalidEnvironmentVariableError('Failed to parse UKEF_GEF_REPORTING_EMAIL_RECIPIENT');
-  }
-};
-
-/**
- * Retrieves the TFM UI URL from the environment variables.
- * @returns The TFM UI URL.
- * @throws {InvalidEnvironmentVariableError} If the TFM_UI_URL environment variable is not defined or is empty.
- */
-export const getTfmUiUrl = (): string => {
-  const { TFM_UI_URL } = process.env;
-
-  if (!TFM_UI_URL) {
-    throw new InvalidEnvironmentVariableError('TFM_UI_URL environment variable is not defined or is empty');
-  }
-
-  return TFM_UI_URL;
-};
 
 /**
  * Sends notification email to UKEF GEF reporting email recipients that a utilisation report has been submitted
